Adds airdrop and faucet functionality
Implements the ability to request an airdrop and create an account
with a faucet.

These functionalities are added to the provider interface and
implemented in both the Maestro and RPC provider implementations.
The Maestro implementation currently throws an error indicating
that it's not yet implemented, while the RPC implementation uses
the postData function to interact with the node.